+++ src/main/kotlin/racingcar/Application.kt
@@ -35,26 +35,6 @@ fun getTrialNumber(): Int {
     return readlnOrNull()?.toIntOrNull() ?: throw IllegalArgumentException("시도할 횟수를 잘못 입력했습니다.")
 }
 
-class Car(val name: String) {
-    var position = 0
-
-    fun move() {
-        if (checkForwardCondition(getRandomNumber())) {
-            position++
-        }
-    }
-
-    private fun getRandomNumber(): Int {
-        val randomNumber = Random.nextInt(0, 10)
-        println(randomNumber)
-        return randomNumber
-    }
-
-    private fun checkForwardCondition(randomNumber: Int): Boolean {
-        return randomNumber >= 4
-    }
-}
-
 fun race(cars: List<Car>, trialNumber: Int) {
     println()
     println("실행 결과")
